Expanded validation code into separate package.
