+++ src/data_validation/tools.py
@@ -1,0 +1,5 @@
+import os
+
+def verify_path(path):
+    is_valid = os.path.exists(path)
+    return is_valid
